Add choice for new or use existing virtual network
- Use one subnet
- Move Virtual Network settings to cluster settings step
- Pass network settings into kibana and jumpbox resources
- Remove "node-" from the naming of vms and allow specifying hostname prefix as a parameter

Host names can be prefixed with up to 5 characters of users control, as cluster unicast discovery uses hostnames to resolve the master nodes (IPs are dynamically assigned).
When deploying a cluster into an existing virtual network, it appears that the Azure network infrastructure will allow a new vm with the same hostname as an existing hostname on the subnet to join an existing network, operating in a "last hostname wins" fashion when it comes to resolution. This results in the new cluster correctly resolving masters from hostnames but is a major issue if a node from the existing cluster is restarted/drops out and rejoins; it will resolve the hostname to the new vm, resulting in bad and unexpected behaviour.

5 character prefix will be enough for all scenarios and would keep the VM naming convention in line with the Azure naming rules (https://azure.microsoft.com/en-us/documentation/articles/guidance-naming-conventions/) for virtual machines that indicate they can be up to 15 characters. A 5 character prefix allows up to a 999 node cluster i.e. {prefix}master-999

Prefix all master, client, data, jumpbox and kibana machines with the specified hostname prefix.