Limit Azure Cloud plugin to using an existing storage account
The usage of nested linked template outputs as parameters to another nested linked template, as well as the use of the listkeys function causes preflight validation checks for core quota limits to not fire within the portal UI. This is an issue within the portal UI that needs to be addressed by Microsoft.

This commit works around the issue by limiting the Azure Cloud plugin to using an existing storage account only and prompting the user to supply the storage account name and access key as input parameters to the template.

- Only allow use of existing storage account for Azure Cloud plugin
- Align README with template parameters and make code examples copy/pasteable
- Add back large deployment tests that trigger core quota limits
- Better descriptive names for Shield configuration
- Group virtual network and azure cloud plugin configuration elements together under sections
- Group node configuration elements together under sections
- Make storage account name and access key required. The validation for the controls being required is only triggered when the controls are visible.