Remove hardcoded Azure locations
Allow a user to specify either ResourceGroup or any Azure location for the the deployment location of the resources in the resource group. Whilst we lose the ability to now validate the passed value against a list of known Azure locations, we gain the flexibility of not needing to update the locations list whenever new Azure locations are rolled out.