Add 6.x versions to the template
This commit adds the 6.x versions of Elasticsearch to the template.
A unqiue bootstrap password is generated within the template and passed to Elasticsearch setup script to set the bootstrap.password in the keystore.
Inverted all 5.x Elasticsearch version checks to use 2.x, since commands are the same across 5.x and and 6.x.