Remove kibana4_server generated password
This commit better aligns the es_kibana user in 2.x and the built-in kibana user in 5.x+.

In 2.x, the es_kibana user is now assigned the kibana4_server role and the kibana4 role
is removed. Kibana will use the es_kibana user credentials when interacting with Elasticsearch.

In 5.x+, the kibana built-in user will use the password supplied for the kibana user and the
es_kibana user is removed. Kibana will use the kibana user credentials when interacting
with Elasticsearch.

This change simplifies the template and a user will now always know the credentials used
by Kibana.

Closes #32