Allow bootstrap.password to be supplied
This commit adds a parameter to allow a user to set the bootstrap.password added to the keystore to set up built-in users.
It can be useful to know this value in the event one needs to re-set up built-in users, and it avoids the need to log in to
each node to set it to a known value.

If not supplied, a unique value is generated within the ARM template, using the uniqueString() function seeded with
the resourceGroup id, deployment name and securityAdminPassword.