Check the elastic user exists
This commit guards against the case where the bootstrap password is the
 same as the elastic superuser password for 5.x and 6.x clusters. In
this scenario, the .security index should also be checked to ensure the
elastic superuser has been created. The _xpack/security/user/elastic
endpoint cannot be used for this because it may return details related
to the bootstrap user.