Improve documentation in portal UI and readme
This commit improves the documentation provided within the portal UI to make it clearer
what each input controls. Remove remnants of account names used in 2.x deployments,
which the template no longer deploys.

Move the built-in user accounts to be the first three within the UI, with es_read user last.

Change naming of Azure Cloud plugin to Azure Repository.