Add support for SAML Single-Sign-On with Azure Active Directory
Push deployment of Kibana public IP address up

This commit pushes the deployment of the Kibana IP public address up
to the same level as the deployment of elasticsearch-nodes.

For SAML Single-Sign-On, the Elasticsearch and Kibana nodes have a dependency on
the public IP address of Kibana when samlServiceProviderUri is not specified as a parameter.

Perform version comparison in separate condition.
Strip trailing slash from port inside echo

Include details about the saml parameters in README

Move Kibana NSG and NIC creation into kibana resources

This commit moves the creation of the Kibana Network Security
Group and Network Interface Card out of the kibana-ip resource
and into the kibana-resources. This allows all ubuntuSettings to
remain in ubuntuSettings.json, including the management port
for connecting to the Kibana VM.

Remove kibana output from ubuntuSettings.json

Add integration test with external loadbalancer and no kibana