Configure Application Gateway to work with TLS on backend pool
This commit adds support for TLS from Application Gateway to the backend pool.
Application Gateway communicates with the backend pool through the internal
loadbalancer, and the public key(s) for certificate(s) used by the backend pool
must be whitelisted by Application Gateway by providing these details to it.

This means that for TLS on the HTTP layer in conjunction with Application Gateway,
only the single esHttpCertBlob option can be supported.

In the case of esHttpCaCertBlob which is used to generate a cert for HTTP layer
for each VM/node in the cluster, the public keys for these certs cannot be
automatically added to Application Gateway as part of deployment.