Pass Elasticsearch HTTP CA cert for Kibana configuration
This commit passes the Elasticsearch HTTP CA cert to Kibana
to configure TLS to Elasticsearch from Kibana. When a CA cert
is provided, it is used to configure the certificate authority.

The present of an Elasticsearch HTTP cert overrides the
presence of a HTTP CA cert. That is, if a HTTP cert is provided,
the certificate authority will be extracted from the PKCS#12 archive.
If no CA cert is present, The verification mode for TLS with Elasticsearch
will be set to none.

Change owner of directory and files in /etc/elasticsearch/ssl