require cert or CA when configuring TLS for HTTP layer
This commit requires that a cert or CA are provided to set up TLS on
the HTTP layer. When a cert is provided, it will take preference, and
Kibana will be configured with certificate verification mode. The most likely
use case for cert is to provide a cert for a CNAME pointing at the external
loadbalancer public IP.

When a CA is provided, HTTP certs are generated for each node, including the
node DNS and IP as Subject Alternative Names, as well as the internal loadbalancer
public IP as a Subject Alternative Name in the cert. Including the internal loadbalancer IP
allows Kibana to be set to full verification mode when communicating internally. A client
communicating through the external loadbalancer can verify certs provided against the CA.
A client communicating through Application Gateway will use the cert configured for the
Gateway.