Add support for SAML Single Sign-On with Azure Active Directory (#188)
Add support for SAML Single Sign-On with Azure Active Directory

This commit adds support for SAML Single Sign-On with Azure Active Directory. Two additional parameters have been added to support this:

samlMetadataUri: the URI from which the metadata for the Identity Provider can be retrieved. 
samlServiceProviderUri: the optional URI of the Service Provider, which would be a CNAME pointing to the Kibana domain name on an Azure sub domain.

- Push deployment of Kibana public IP address up

    the Elasticsearch and Kibana nodes now have a dependency on
the output from the public IP address of Kibana resource, when samlServiceProviderUri is not specified as a parameter.

Closes #179