Patch marketplace tracking guid in gulp pipe transform
This commit always changes the tracking guid in the gulp pipe
to use the Marketplace guid, so that this guid always ends up in the
release zip. This ensures that the guid always makes it into the zip
and can never be accidently checked into GitHub with this guid applied.