Use resource group to get details for azure-repository storage account
This commit updates the parameters required to configure azure repository plugin with an
existing storage account in another resource group. Instead of passing the storage account
key and endpoint suffix, these values can be determined with the storage account name and
resource group.

key and suffix are removed and replaced with resourceGroup.

Note that this now requires that the storage account and resoruce group exist in the same
subscription. The implementation could be expanded in future to accept a subscriptionId,
if necessary.

Closes #215