Enable accelerated networking
This commit enables accelerated networking by default for VM SKUs that support it,
for both Elasticsearch nodes and the Kibana VM.

Exposes the option to override accelerated networking if necessary. This is exposed in
the case that new VMs are introduced that support accelerated networking,
or existing VMs are updated to enable accelerated networking, giving users the ability to
control accelerated networking behaviour.

Closes #224