Allow bash special characters to be passed in input parameters (#231)
* Use single quotes to pass script parameters

This commit changes escaping parameter values passed to shell
scripts with double quotes to single quotes. Use a variable for single quotes;
whilst single quotes can be escaped by single quotes, using a variable looks
a little cleaner.

* Include special characters in YAML configuration

This commit updates the YAML configuration for Kibana to include special characters.
The double quote needs to be escaped inside of the JSON string, and the backslash needs to be
doubly escaped; once to escape within the shell script parameter and again to escape
the value written to yaml configuration. The resulting shell script parameter is

    'server.name: \"My strangely named server $`'\"'\"'\\\"~;\"\nserver.defaultRoute: \"/app/kibana\"'

which is written to yaml as

    server.name: "My strangely named server $`'\"~;"
    server.defaultRoute: "/app/kibana"

Closes #222