Check sha1 checksum for versions less than 5.6.2 (#241)
This commit updates the install scripts to download and use a
sha1 checksum for Elastic Stack versions less than 5.6.2.

Fixes #237

* Include package name in sha file

This commit adds the package name to the sha checksum
file, to validate it with shasum utility. Earlier versions
of checksum file only include the checksum and not the package name.

* Use esVersion for parameter instead of version

With the update to gulp, the version parameter is now used by
gulp to return details of the gulp version. Change our version
command line flag to esVersion, in line with the template parameters

* don't delete resource groups on bailout when nodestroy passed

This commit changes the behaviour of bailout to not delete resource groups
when nodestroy flag has been passed. When there's a failure that causes a
bailout, user may want to investigate by passing --nodestroy.