Add certificates for Elasticsearch and Kibana TLS to UI definition
This commit adds the ability to supply certificates for Transport Layer Security for both Elasticsearch and
Kibana.

With Elasticsearch 6.x and version >= 6.8.0, and 7.x and version >= 7.1.0, TLS can be configured for basic license level.

Since deploying through Azure Marketplace (and hence using createUiDefinition) always deploys with a trial license, it is safe to always provide the option to configure SSL/TLS certificates.