Add support for Elastic Stack 7.1.1 (#281)
This commit adds support for Elastic Stack 7.1.1 and 6.8.0. It includes a change to configure basic security when

xPackPlugins is 'Yes'

or

Version is >= 6.8.0 and < 7.0.0, or >= 7.1.0

and configures TLS when certificates have been supplied,

Add an integration test for basic license with TLS when the version is >= 7.1.0. This test configures TLS on the HTTP and Transport layers and asserts both Kibana and Logstash can communicate with the cluster.

Closes #279