Remove usage of outputs and inline functions
This commit removes the usage of output values for the new or existing storage account
for the azure repository plugin, instead calculating the values in line in mainTemplate.json.

Similarly, remove the usage of the output value of kibana-ip as an input to virtual-machines, instead
calculating in line.

These changes are to allow preflight validation checks to *not* be short circuited for the virtual-machines resource,
so that a deployment will fail early when there are insufficient quotas to deploy the requested resources. This is
to workaround a limitation in ARM templates that short circuits preflight validation checks for a linked template when
it depends on the output of another.