Updated expected validation of large clusters
This commit updates the validation assertion for large cluster integration tests.

Preflight validation checks performed by the Azure infrastructure are currently
broken because of the usage pattern of using linked template outputs as inputs
into another linked template. This causes preflight validation checks to not
be performed on any resources in the latter linked template. There appears no
easy way to remedy this because of limitations with the available template
functions.

For now, assert that the large cluster integration tests will pass validation.