Allow multiple instances of Logstash
This commit allows multiple instances of Logstash to be deployed. Similar to data, client and master nodes, multiple Logstash VMs are deployed in an Availability set, with fault and update domains.

The naming of logstash resources follow the convention for multiple counts of a resource i.e. logstash-0, logstash-1, etc. prefixed with the vmHostNamePrefix.

Closes #247