Set trial license only when INSTALL_XPACK = Yes and >= 6.3.0  (#304)
This commit sets the trial license in elasticsearch.yml only when INSTALL_XPACK=1 and
the version is greater than or equal to 6.3.0.

Fixes a bug where trial license was applied when BASIC_SECURITY=1 and the version
is greater than or equal to 6.3.0.

Other changes:

- check the license applied to the deployment using the Elasticsearch APIs, if external load balancer or application gateway is deployed.
- emit total, passed and failed tests on command line for test runs
- use the same timestamp in integration test resource groups
- deploy an NSG for the backend pool when using a Standard SKU external load balancer

Fixes #303

(cherry picked from commit 8903dccca396b7fb5da434979db228fde30693d1)