Calculate vNet load balancer IP
This commit changes the output from createUiDefintion to calculate the IP for the internal load balancer from the start address of the subnet. With the move to using Microsoft.Azure.CreateUIDef, the VirtualNetworkCombo UI element now returns the start address of the subnet, *not* the first usable starting address like it used to e.g. returns 10.0.0.1 and not 10.0.0.4

Azure reserves 5 IP addresses within each subnet, x.x.x.0-x.x.x.3 and the last address: https://docs.microsoft.com/en-us/azure/virtual-network/virtual-networks-faq#are-there-any-restrictions-on-using-ip-addresses-within-these-subnets

There is no UI function to add to an IP address, so perform through string splitting and concatenation.