Move VNetLoadBalancerIp address calculation to mainTemplate.json
Relates: https://github.com/MicrosoftDocs/azure-docs/issues/37958

This commit moves the calculation of a start address for a new subnet from createUiDefinition.json to mainTemplate.json.

When a new vnet is selected, createUiDefinition.json returns a reserved IP address as the startAddress for the new subnet.
The starting reserved addresses for a new subnet should be calculated, along with the first available address. In the event
that the value passed for vNetLoadBalancerIp is a reserved address (as will be the case for a new subnet from the UI), use the
calculated first address as the value for vNetLoadBalancerIp.