Update default VM sizes
Supersedes: #276,  #320

This commit updates the default VM sizes for Elasticsearch, Logstash and Kibana.

Remove the D-series and replace the default Standard_D1 with Standard_DS1_v2.
Remove the A-series and replace the Standard_A2 with Standard_A2_v2

Add Av2-series
Add Easv4-series
Add Dasv4-series
Remove DS-series

Update the VM sizes in the documentation, and add a Q&A to the FAQs
for when to use the Marketplace offering vs. the ARM template directly.