Deploy NSG and VM public IPs when internal standard LB selected (#325)
This commit deploys a Network Security Group when a
Standard SKU internal Load Balancer is selected and the
loadBalancerType is internal. 

Each VM in the backend pool is also assigned a Standard 
SKU public IP address to allow outbound traffic. Different 
Basic and Standard SKU load balancer or public IP resources 
in an Availability Set is not allowed, so public IPs 
must be Standard SKU for VMs when deploying standard 
internal loadbalancer.

Add integration tests for Standard Internal load balancer when
used with loadBalancerType `internal` and `gateway`

Fixes #308