Remove superfluous function
This commit renames _curl_with_error_code
to curl_ignore_409 and removes the original
curl_ignore_409 function.

When curl_ignore_409 is called,
the status code is checked and
deployment fails when the status code
is not 0. This was not being respected as
intended, because the status code being
returned is that of running sed command.