Retry waiting for security index to be green
Supersedes: #372

Add retries around waiting for the security index to be
green, to mitigate a race condition:

1. node A updates elastic user password using the bootstrap password to authenticate
2. node B tries to update the elastic user password using the bootstrap password, but fails, Expected failures are that
	1. node A change is already reflected and a 401 response is returned
	2. security index is not yet green, and an unavailable_shards_exception 503 response is returned
3. node B tries to ping the cluster the cluster with the updated elastic user password

It looks like the expected failure in 2.2 would fail immediately with a 401 response because
the security index isn't yet ready, meaning the cluster expects the bootstrap password to authenticate.
But, at the point the security index is ready, the cluster would return a 401 response when using
the bootstrap password and return the status when using the updated password.

Closes #363